Changed name of file and updated user
